requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.3.3
bandit==1.6.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==3.0.6
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
djangorestframework==3.11.0
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
logilab-common==1.6.4
mccabe==0.6.1
oauthlib==3.1.0
pbr==5.4.5
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
PySocks==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
stevedore==1.32.0
tweepy==3.8.0
urllib3==1.25.9
virtualenv==20.0.21
Werkzeug==1.0.1
wrapt==1.11.2
